+++ SymonServer/Tcp/TcpStream.cs
@@ -79,7 +79,7 @@ namespace Symon.Server {
                     Array.Copy(buffer, recvData, recv);
                     recvList.AddRange(recvData);
                     while (true) {
-                        if (length == 0) {
+                        if (length == 0 && recvList.Count >= 4) {
                             length = BitConverter.ToInt32(recvList.ToArray(), 0);
                             recvList.RemoveRange(0, 4);
                         }
